@@ -1,17 +1,17 @@
 package com.gabrielbmoro.programmingchallenge.ui.homescreen.pages
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ProgressBar
+import androidx.databinding.DataBindingUtil
 import com.gabrielbmoro.programmingchallenge.ProgrammingChallengeApp
 import com.gabrielbmoro.programmingchallenge.R
 import com.gabrielbmoro.programmingchallenge.api.ApiServiceAccess
 import com.gabrielbmoro.programmingchallenge.dao.FavoriteMovieDAOAssistant
+import com.gabrielbmoro.programmingchallenge.databinding.FragmentTopRatedMoviesBinding
 import com.gabrielbmoro.programmingchallenge.models.Movie
 import com.gabrielbmoro.programmingchallenge.ui.CellSimpleMovieAdapter
 import rx.android.schedulers.AndroidSchedulers
@@ -34,17 +34,18 @@ interface TopRatedPageContract {
     interface Presenter {
         fun loadMovies()
     }
+
     /**
      * View defines the setup and update screen operations.
      * @author Gabriel Moro
      * @since 2018-08-30
      */
-    interface View{
+    interface View {
         fun setupRecyclerView()
-        fun onNotifyDataChanged(moviesList : ArrayList<Movie>)
+        fun onNotifyDataChanged(moviesList: ArrayList<Movie>)
         fun showProgress()
         fun hideProgress()
-        fun areThereSomeElementsAtRecyclerView() : Boolean
+        fun areThereSomeElementsAtRecyclerView(): Boolean
     }
 }
 
@@ -53,19 +54,19 @@ interface TopRatedPageContract {
  * @author Gabriel Moro
  * @since 2018-08-30
  */
-class TopRatedFragment : Fragment(), TopRatedPageContract.View {
+class TopRatedFragment : androidx.fragment.app.Fragment(), TopRatedPageContract.View {
 
-    private var presenter       : TopRatedPageContract.Presenter? = null
-    private var mrvRecyclerView : RecyclerView? = null
-    private var mpdProgressBar  : ProgressBar?  = null
+    private lateinit var binding : FragmentTopRatedMoviesBinding
+    private var presenter: TopRatedPageContract.Presenter? = null
 
     /**
      * In this state the fragment view is created.
      * @author Gabriel Moro
      * @since 2018-08-30
      */
-    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater?.inflate(R.layout.fragment_top_rated_movies, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_top_rated_movies, container, false)
+        return binding.root
     }
 
     /**
@@ -74,11 +75,9 @@ class TopRatedFragment : Fragment(), TopRatedPageContract.View {
      * @author Gabriel Moro
      * @since 2018-08-30
      */
-    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        mrvRecyclerView = view?.findViewById(R.id.rvTopRatedMoviesList)
-        mpdProgressBar = view?.findViewById(R.id.pbProgress)
-        mpdProgressBar?.isIndeterminate = true
+        binding.pbProgress.isIndeterminate = true
         setupRecyclerView()
         presenter = TopRatedPresenter(this)
     }
@@ -100,8 +99,8 @@ class TopRatedFragment : Fragment(), TopRatedPageContract.View {
      */
     override fun setupRecyclerView() {
         val llManager = LinearLayoutManager(context)
-        mrvRecyclerView?.layoutManager = llManager
-        mrvRecyclerView?.adapter = CellSimpleMovieAdapter(ArrayList())
+        binding.rvTopRatedMoviesList.layoutManager = llManager
+        binding.rvTopRatedMoviesList.adapter = CellSimpleMovieAdapter(ArrayList())
     }
 
     /**
@@ -109,23 +108,23 @@ class TopRatedFragment : Fragment(), TopRatedPageContract.View {
      * @author Gabriel Moro
      * @since 2018-08-30
      */
-    override fun onNotifyDataChanged(amoviesList : ArrayList<Movie>) {
-        (mrvRecyclerView?.adapter as CellSimpleMovieAdapter).mlstMovies = ArrayList(amoviesList)
-        mrvRecyclerView?.adapter?.notifyDataSetChanged()
+    override fun onNotifyDataChanged(moviesList: ArrayList<Movie>) {
+        (binding.rvTopRatedMoviesList.adapter as CellSimpleMovieAdapter).mlstMovies = ArrayList(moviesList)
+        binding.rvTopRatedMoviesList.adapter?.notifyDataSetChanged()
     }
 
     override fun hideProgress() {
-        mpdProgressBar?.visibility = ProgressBar.GONE
+        binding.pbProgress.visibility = ProgressBar.GONE
     }
 
     override fun showProgress() {
-        mpdProgressBar?.visibility = ProgressBar.VISIBLE
+        binding.pbProgress.visibility = ProgressBar.VISIBLE
     }
 
     override fun areThereSomeElementsAtRecyclerView(): Boolean {
-        return if(mrvRecyclerView?.adapter == null) false
+        return if (binding.rvTopRatedMoviesList.adapter == null) false
         else {
-            val cellSimpleAdapter = (mrvRecyclerView?.adapter as CellSimpleMovieAdapter)
+            val cellSimpleAdapter = (binding.rvTopRatedMoviesList.adapter as CellSimpleMovieAdapter)
             cellSimpleAdapter.mlstMovies.isNotEmpty()
         }
     }
@@ -136,10 +135,10 @@ class TopRatedFragment : Fragment(), TopRatedPageContract.View {
  * @author Gabriel Moro
  * @since 2018-08-30
  */
-class TopRatedPresenter(aview : TopRatedPageContract.View) : TopRatedPageContract.Presenter {
+class TopRatedPresenter(aview: TopRatedPageContract.View) : TopRatedPageContract.Presenter {
 
-    private val view               : TopRatedPageContract.View = aview
-            var mlstMoviesFiltered : ArrayList<Movie>? = null
+    private val view: TopRatedPageContract.View = aview
+    private var mlstMoviesFiltered: ArrayList<Movie>? = null
 
     /**
      * This method allows the load movies using the request.
@@ -149,8 +148,8 @@ class TopRatedPresenter(aview : TopRatedPageContract.View) : TopRatedPageContrac
      * @since 2018-08-30
      */
     override fun loadMovies() {
-        if(!ProgrammingChallengeApp.mbHasNetworkConnection) return
-        if(view.areThereSomeElementsAtRecyclerView()) return
+        if (!ProgrammingChallengeApp.mbHasNetworkConnection) return
+        if (view.areThereSomeElementsAtRecyclerView()) return
         view.showProgress()
         val api = ApiServiceAccess()
         api.getMovies("vote_average.desc")
@@ -158,21 +157,21 @@ class TopRatedPresenter(aview : TopRatedPageContract.View) : TopRatedPageContrac
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(
                         {
-                            if(it.mlstResults != null) mlstMoviesFiltered = ArrayList(it.mlstResults!!)
+                            if (it.results != null) mlstMoviesFiltered = ArrayList(it.results!!)
 
                             val daoAssistant = FavoriteMovieDAOAssistant(ProgrammingChallengeApp.mappDataBuilder!!)
-                            mlstMoviesFiltered?.forEach {
-                                it.mstrPosterPath = "https://image.tmdb.org/t/p/w154"+ it.mstrPosterPath
-                                if(ProgrammingChallengeApp.mappDataBuilder != null) {
-                                    it.mbIsFavorite = daoAssistant.isThereSomeMovieInFavorite(it)
+                            mlstMoviesFiltered?.forEach { movie ->
+                                movie.posterPath = "https://image.tmdb.org/t/p/w154" + movie.posterPath
+                                if (ProgrammingChallengeApp.mappDataBuilder != null) {
+                                    movie.isFavorite = daoAssistant.isThereSomeMovieInFavorite(movie)
                                 }
                             }
                         },
                         {
                             it.printStackTrace()
-                        },{
+                        }, {
                     //ui
-                    if(mlstMoviesFiltered!=null) {
+                    if (mlstMoviesFiltered != null) {
                         view.onNotifyDataChanged(mlstMoviesFiltered!!)
                         view.hideProgress()
                     }
